requirements.txt
backoff==2.2.1
grpcio==1.50.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-storage==2.5.0
google-cloud-dataproc==5.4.1
flask~=2.3.2
joblib~=1.2.0
nltk~=3.8.1
pandas~=2.0.1
scikit-learn~=1.2.2
numpy
beautifulsoup4~=4.12.2
pyspark~=3.4.0
scipy~=1.10.1